Fixed not compiling on Windows.

+++ sources/gui/RAGE_scene_view.cpp
@@ -42,8 +42,8 @@ void RAGE_scene_view::draw(RAGE *rage)
 	glBindFramebuffer(GL_FRAMEBUFFER, 0);
 
 	ImGui::Begin("Scene", NULL, ImGuiWindowFlags_NoCollapse);
-	ImGui::SetWindowPos(ImVec2(rage->window->pixel_position.x, rage->window->pixel_position.y), ImGuiCond_FirstUseEver);
-	ImGui::SetWindowSize(ImVec2(this->window_size.x, this->window_size.y), ImGuiCond_FirstUseEver);
+	ImGui::SetWindowPos(ImVec2((float)rage->window->pixel_position.x, (float)rage->window->pixel_position.y), ImGuiCond_FirstUseEver);
+	ImGui::SetWindowSize(ImVec2((float)this->window_size.x, (float)this->window_size.y), ImGuiCond_FirstUseEver);
 	rage->camera.set_aspect_ratio(this->window_size);
 	this->window_size = glm::ivec2(ImGui::GetWindowSize().x, ImGui::GetWindowSize().y);
 	ImGui::Text("Window size: %d, %d", this->window_size.x, this->window_size.y);
